Combine modules asyn_lpa and label_propagation.
add a print statement for debugging strange py3.4.4 32-bit error
