+++ networkx/algorithms/community/__init__.py
@@ -15,7 +15,6 @@ functions as attributes of ``community``. For example::
     [[0, 1, 2, 3, 4], [5], [6, 7, 8, 9, 10]]
 
 """
-from networkx.algorithms.community.asyn_lpa import *
 from networkx.algorithms.community.asyn_fluidc import *
 from networkx.algorithms.community.centrality import *
 from networkx.algorithms.community.community_generators import *
